+++ example/lib/main.dart
@@ -24,13 +24,19 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   void initState() {
     super.initState();
-    String DNASequence = generateRandomNucleotideSequence(40, 'DNA');
-    String RNASequence = generateRandomNucleotideSequence(40, 'RNA');
+
+    String DNASequence = makeNucleotideSequence(40, 'DNA');
+    String RNASequence = makeNucleotideSequence(40, 'RNA');
     print(validateNucleotideSequence(DNASequence, 'DNA'));
-    print(countNucleotideFrequency(DNASequence, 'DNA'));
-    print(transcribeDNAToRNA('GATGGAACTTGACTACGTAAATT'));
-    print(generateComplementaryStrand('AAAACCCGGT', 'DNA', true));
-    print(calculateGCContent('CCACCCTCGTGGTAGGCAGTAGGTGGAAT', 'DNA'));
+    print(nucleotideFrequency(DNASequence, 'DNA'));
+    print(DNAToRNA('GATGGAACTTGACTACGTAAATT'));
+    print(complementaryStrand('AAAACCCGGT', 'DNA', true));
+    print(GCContent('CCACCCTCGTGGTAGGCAGTAGGTGGAAT', 'DNA'));
+    print(hammingDistance('GAGCCTACTAACGGGAT', 'CATCGTAATGACGGCCT', 'DNA'));
+    print(dominantPhenotypeProba(19, 29, 28)); // 0.6892982456140351
+    print(detectSequenceType('AGCGG'));
+    print(RNAtoProtein('AUGGCCAUGGCGCCCAGAACUGAGAUCAAUAGUACCCGUAUUAACGGGUGA'));
+    print(matchMotif('GATATATGCATATACTT', 'ATAT', 'DNA'));
   }
 
   @override