+++ lib/constants/lists.dart
@@ -1,0 +1,3 @@
+const List<String> DNANucleotides = ["A", "T", "G", "C"];
+const List<String> RNANucleotides = ["A", "U", "G", "C"];
+const List<String> translationStopCodons = ["UGA", "UAA", "UAG"];