+++ lib/constants/strings.dart
@@ -1,0 +1,18 @@
+const String kDNA = 'dna';
+const String kRNA = 'rna';
+const String kInvalidSequence = 'invalidSequence';
+const String kAmbiguousSequence = 'ambiguousSequence';
+const String kResult = 'result';
+const String kInvalidRNAErrorMessage = 'invalidRNAErrorMessage';
+const String kInvalidDNAErrorMessage = 'invalidDNAErrorMessage';
+const String kTranslationStopCodon = 'translationStopCodon';
+const String kPolypeptideSequence = 'polypeptideSequence';
+const String kRNASequenceNucleotideCount = 'RNASequenceNucleotideCount';
+const String kPolypeptideSequenceAminoAcidCount =
+    'polypeptideSequenceAminoAcidCount';
+const String kStartIndex = 'startIndex';
+const String kEndIndex = 'endIndex';
+const String kMatchCount = 'matchCount';
+const String kMatchIndices = 'matchIndices';
+const String kSequence = 'sequence';
+const String kSequenceType = 'type';