+++ lib/errors/invalidSequenceTypeErrorMessage.dart
@@ -1,0 +1,3 @@
+String invalidSequenceTypeErrorMessage(String sequenceType) {
+  return "Invalid Sequence Type Error. '$sequenceType' is not a valid sequence type. Please enter the argument 'DNA' or 'RNA' for parameter 'sequenceType'.";
+}