+++ lib/helpers/detectStartCodons.dart
@@ -1,0 +1,1 @@
+detectStartCodons(String nucleotideSequence) {}