+++ lib/helpers/loadFASTA.dart
@@ -1,0 +1,1 @@
+loadFASTA(String path) {}