+++ lib/helpers/reverseNucleotideSequence.dart
@@ -1,0 +1,3 @@
+String reverseNucleotideSequence(String nucleotideSequence) {
+  return nucleotideSequence.split('').reversed.join('');
+}