+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.3] - 2021-03-01
+
+* Added hammingDistance function.
+* Added matchMotif function.
+* Added RNAtoProtein function. 
+* Added dominantPhenotypeProba function.
+* Shortened existing function names.
+
 ## [0.0.2] - 2021-02-22
 
 * Added validateNucleotideSequence function.